requirements.txt
pytest
requests
allure-pytest
curlify
jsonschema
